requirements.txt
transformers
albumentations
opencv-python-headless
boto3
Pillow
requests
numpy
tqdm
wandb